Adds "blobify" as auxiliary function
I found I was using it almost in every test, so I added it, with
tests (one) to the main module. I understand this is major, since it
adds stuff to the interface (and you might not like the name) so I'll
wait until this is accepted, reviewed or rejected to continue. As it
is now, it would be very useful to refactor a good part of the
tests...
Refs #5